incorporated comments, moved findWorkspaceDirByName function to helpers.ts, fixed all the comments
