+++ package.json
@@ -30,6 +30,7 @@
     "@storybook/react": "^10.1.10",
     "@storybook/react-webpack5": "^10.1.10",
     "@stylistic/eslint-plugin": "^5.3.1",
+    "@types/adm-zip": "^0.5.8",
     "@types/jest": "^29.5.11",
     "@types/js-yaml": "^4.0.9",
     "@types/lodash-es": "^4.17.12",
@@ -105,6 +106,7 @@
     }
   },
   "dependencies": {
+    "adm-zip": "^0.5.17",
     "ajv": "^8.17.1",
     "git-url-parse": "^14.1.0"
   }
