+++ tests/utils/helpers.ts
@@ -1,4 +1,6 @@
 import fs from 'fs';
+import yaml from 'js-yaml';
+import path from 'path';
 
 /**
  * Read and JSON-parse a file, retrying until it has parseable content.
@@ -22,4 +24,19 @@ export async function readExportedJson(filePath: string, { tries = 50, interval 
     await new Promise((resolve) => setTimeout(resolve, interval));
   }
   throw lastError ?? new Error(`Timed out waiting for valid JSON at ${filePath}`);
+}
+
+export type WorkspaceConfig = {
+  info?: { name: string; type: string };
+};
+
+/** Find a workspace dir under `location` whose workspace.yml has the given name. */
+export function findWorkspaceDirByName(location: string, name: string): string | null {
+  for (const entry of fs.readdirSync(location)) {
+    const ymlPath = path.join(location, entry, 'workspace.yml');
+    if (!fs.existsSync(ymlPath)) continue;
+    const config = yaml.load(fs.readFileSync(ymlPath, 'utf8')) as WorkspaceConfig;
+    if (config?.info?.name === name) return path.join(location, entry);
+  }
+  return null;
 }
